# admin/Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BarberSangaji Backend API
  version: "1.1.0"
  description: >
    API untuk BarberSangaji — autentikasi, upload file menggunakan UploadThing, konten CRUD, dan testimonial.
    Mendukung upload file langsung lewat `multipart/form-data` pada endpoint konten & testimonial.
servers:
  - url: http://localhost:3000/api
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Content:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [foto, video]
        file_name:
          type: string
        file_path:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, title, description, category, file_name, file_path]
    Testimonial:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        profession:
          type: string
        content:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        photo_path:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, name, profession, content, rating]
...
paths:
...
  /content:
    get:
      summary: List konten (foto/video)
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [foto, video]
      responses:
        '200':
          description: Daftar konten
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Content'

    post:
      summary: Upload dan buat konten baru
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                  enum: [foto, video]
                file:
                  type: string
                  format: binary
              required: [title, description, category, file]
      responses:
        '201':
          description: Konten berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          description: Missing required fields
...
  /testimonials:
    get:
      summary: List semua testimonial
      responses:
        '200':
          description: Array of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Testimonial'

    post:
      summary: Upload dan buat testimonial
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string }
                profession: { type: string }
                content: { type: string }
                rating: { type: integer, minimum: 1, maximum: 5 }
                file:
                  type: string
                  format: binary
              required: [name, profession, content, rating]
      responses:
        '201':
          description: Testimonial berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testimonial'
...